Refactor serverLogger
We will now remove the ambiguity whether `error` levels should be logged to Sentry.
All "errors" will now be logged to Sentry. This should mean that if you
use `serverLogger` in a catch block (or wherever).

This means that it is your intent to see this in the "server's logs"
when that exception is caught. This also prevents us from "error swallowing"
which can lead to unknown bugs.

Simple logging can still be achieved by doing the following:
```
  serverLogger(message, 'log')
```

`serverLogger` will also now support capturing contexts similar
to what Sentry.captureException does. By exposing this to the dev,
we are providing a way for them to provide contexts similar to how
Sentry expects it instead of converting JSON errors to strings

https://docs.sentry.io/platforms/node/enriching-events/context/#passing-context-directly